requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
requests==2.31.0
psycopg[binary]==3.3.3
kubernetes==28.1.0
psutil==5.9.8
sqlalchemy==2.0.23
aiosqlite==0.20.0
prometheus-client==0.20.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-fastapi==0.44b0
pyyaml==6.0.1
dependency-injector==4.41.0
redis==5.0.1
networkx==3.2.1
PyJWT==2.8.028
